refactor/#372 모달 분리
